refactor dynamic_config: move public testsuite classes from pytest plugin commit_hash:9c994de87e27a55450a990914265d4cc577d8658
