+++ .mapping.json
@@ -4234,6 +4234,7 @@
   "testsuite/pytest_plugins/pytest_userver/__init__.py":"taxi/uservices/userver/testsuite/pytest_plugins/pytest_userver/__init__.py",
   "testsuite/pytest_plugins/pytest_userver/chaos.py":"taxi/uservices/userver/testsuite/pytest_plugins/pytest_userver/chaos.py",
   "testsuite/pytest_plugins/pytest_userver/client.py":"taxi/uservices/userver/testsuite/pytest_plugins/pytest_userver/client.py",
+  "testsuite/pytest_plugins/pytest_userver/dynconf.py":"taxi/uservices/userver/testsuite/pytest_plugins/pytest_userver/dynconf.py",
   "testsuite/pytest_plugins/pytest_userver/grpc/__init__.py":"taxi/uservices/userver/testsuite/pytest_plugins/pytest_userver/grpc/__init__.py",
   "testsuite/pytest_plugins/pytest_userver/grpc/_mocked_errors.py":"taxi/uservices/userver/testsuite/pytest_plugins/pytest_userver/grpc/_mocked_errors.py",
   "testsuite/pytest_plugins/pytest_userver/grpc/_reflection.py":"taxi/uservices/userver/testsuite/pytest_plugins/pytest_userver/grpc/_reflection.py",
