+++ core/functional_tests/dynamic_configs/tests_kill_switches/constants.py
@@ -1,8 +1,8 @@
 import pytest
-from pytest_userver.plugins import dynamic_config as dynamic_config_lib
+from pytest_userver import dynconf
 
 CONFIG_KEY = 'TEST_CONFIG_KEY'
 
 NEW_VALUE = 'new_value'
 
-KILL_SWITCH_DISABLED_MARK = pytest.mark.config(**{CONFIG_KEY: dynamic_config_lib.USE_STATIC_DEFAULT})
+KILL_SWITCH_DISABLED_MARK = pytest.mark.config(**{CONFIG_KEY: dynconf.USE_STATIC_DEFAULT})
